requirements.txt
mistralai
google-generativeai
pandas
python-dotenv
pillow
requests
psycopg2-binary
